Fix imports and remove unnecessary tests
